+++ src/debug/symbols.h
@@ -30,10 +30,10 @@ extern bool Symbols_GetCpuAddress(symtype_t symtype, const char *name, Uint32 *a
 extern bool Symbols_GetDspAddress(symtype_t symtype, const char *name, Uint32 *addr);
 /* symbol address -> name search */
 extern const char* Symbols_GetByCpuAddress(Uint32 addr);
-extern const char* Symbols_GetByDspAddress(Uint16 addr);
+extern const char* Symbols_GetByDspAddress(Uint32 addr);
 /* symbol address -> index */
 extern int Symbols_GetCpuAddressIndex(Uint32 addr);
-extern int Symbols_GetDspAddressIndex(Uint16 addr);
+extern int Symbols_GetDspAddressIndex(Uint32 addr);
 /* how many symbols are loaded */
 extern int Symbols_CpuCount(void);
 extern int Symbols_DspCount(void);
